djangoapp/requirements.txt
asgiref==3.7.2
Django==5.0.2,<5.1
Faker==23.1.0
psycopg2-binary==2.9.9,<2.10
python-dateutil==2.8.2
six==1.16.0
sqlparse==0.4.4
pillow==10.2.0
